feat(client): modularize schemas and add linting script
This commit introduces a major refactoring to improve code structure and adds a dedicated linting script for the client application.

- Moves Zod schemas from page components into a dedicated `@/schemas` directory, creating a single source of truth for validations.
- Standardizes import order and formatting across several UI components for better code consistency.
- Improves default form values and the handling of optional fields in event-related forms to prevent bugs.
- Adds a new `lint:client` script to the root `package.json` to enable isolated client-side linting.
