+++ package.json
@@ -13,6 +13,7 @@
     "dev:server": "yarn workspace @checkfacil/server start",
     "dev": "concurrently \"yarn dev:client\" \"yarn dev:server\"",
     "lint": "eslint . --max-warnings 0",
+    "lint:client": "eslint client/src --max-warnings 0",
     "lint:fix": "eslint . --fix --max-warnings 0",
     "format": "prettier --write \"**/*.{ts,tsx,js,mjs,json,md,css}\"",
     "format:check": "prettier --check \"**/*.{ts,tsx,js,mjs,json,md,css}\"",
